# apps/backend/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "backend API",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/ingress/credentials": {
      "get": {
        "tags": [
          "admin"
        ],
        "operationId": "get_ingress_credentials",
        "responses": {
          "200": {
            "description": "Whether ingress credentials are configured per platform",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngressCredentialsResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "operationId": "revoke_ingress_credentials",
        "parameters": [
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/IngressPlatform"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Ingress credentials revoked"
          },
          "400": {
            "description": "Platform not configured"
          },
          "500": {
            "description": "Failed to clear credentials"
          }
        }
      }
    },
...
    "/wapi/queue/{id}/complete": {
      "post": {
        "tags": [
          "queue"
        ],
        "operationId": "complete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/QueueEntryId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Spin completed"
          },
          "404": {
            "description": "Queue entry not found"
          },
          "409": {
            "description": "Entry is not in spinning state"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "IngressCredentialsResponse": {
        "type": "object",
        "required": [
          "twitch",
          "vk_video_live"
        ],
        "properties": {
          "twitch": {
            "type": "boolean"
          },
          "vk_video_live": {
            "type": "boolean"
          }
        }
      },
...
      "QueueEntryId": {
        "type": "integer",
        "format": "int32",
        "minimum": 0
      },
...
    }
  },
...
}